Moving out gc related queue code to another module(cls_rgw_queue).
